(#1764) Add test for new enhanced exit code
Now, when you have useEnhancedExitCodes turned on, if you try to
add an already existing source, remove a source that doesn't exist,
disable a source that is already disabled, or enable a source that is
already enabled, the exit code will change to be a 2, giving a clear
indication that no action was taken.

This commit adds a Pester test to verify that this works as expected.