(#23) Add tests to handle install all
This commit adds some Pester tests to ensure that any attempts to run
the choco install all command is correctly handled when targeting
an alternative source.  This is simply asserting that there is an
error exit code, and that the correct output is displayed.

Some of the new tests are specifically set to only run on Test-Kitchen,
as there is the potential for changes to happen on the host machine,
and we don't want to impact people.