(test) Update assertions for push commands
Something has changed in the output that is coming back from the CCR
when pushing a package, and that package fails. It is not immediately
clear what has changed, but the assertions are being updated to allow
the tests to pass, and we can monitor this going forward to see if
something changes again.