(#2050) Add Pester test for --ignore-pinned option
With the next release of Chocolatey CLI, it is possible to run choco
upgrade --ignore-pinned, rather than first having to run a choco pin
remove command.

This commit adds a Pester test to cover the usage of the
--ignore-pinned option on the upgrade command.  This tests relies on
correct execution of the --pin option on the install command, but this
is tested elsewhere.