(#310) Check for presence of Location before usage
During testing, it was found that the Location property of the Assembly
was an entry string, therefore the cmdlets weren't being loaded.

This commit attempts to address this by first checking the Location
property, and if it is there, use it.  Otherwise, fall back to the
previous direct Import-Module against the $licensedAssembly.

Will schedule a call with Rain to discuss this further once they are
back from holiday.