(#310) Ensure correct loading of profile
During testing, it was found that the Chocolatey Profile was not being
correctly loaded, with neither  Update-SessionEnvironment nor
refreshenv working as expected.

This commit changes the profile loading to explicitly set the alas for
refreshenv, and also to load Update-SessionEnvironment as a Cmdlet,
rather than a Funcation.  This is needed since it is now written in C#
rather than PowerShell.