(doc) Change source code to match production docs
While running the GenerateDocs.ps1 script, ahead of releasing version
2.3.0, it was found that the production docs has been updated to make
some changes, for example, removal of older version numbers that no
longer need to be displayed.  This commit addresses this change in the
production docs, but making the same change in the source code, so that
the next time that GenerateDocs.ps1 is run, the output will better
match what is already in place.