(build) Make Agent requirements more specific
It was found that some builds are failing due to picking up the wrong
build agent, and as a result, some required applications were not
installed, so the build failed.

This commit makes the Build Agent requirements more specific to ensure
that the Agent Name is taken into consideration.  This was done as an
additional check, rather than replacing the existing requirements, as
they are simply being more explicit about the requirement, the existing
requirements are still valid.