(tests) Remove v2 import tests
PowerShell V2 is no longer supported. With the change to compiled
cmdlets, these will not work on any system that doesn't have PowerShell
v3 or newer due to the .Net Framework used.

This removes the tests as they will be unable to load the compiled
cmdlets and fail the tests.