(#1901) Revert commit to display location
Based on some testing that was completed, we can't bring in this change
currently, since we can't provide the same functionality in CLE without
bumping the version number of the CLI dependency in the nuspec file.
This is not something that we are ready to do, so taking out this
change, and the issue will be moved to the next milestone.