(#23) Move call to SetRemotePackageNamesIfAllSpecified
The SetRemotePackageNamesIfAllSpecified method HAS to be called AFTER
the call to GetRemoteRepositories.

In that method, the Sources on the configuration object are expanded to
be the full source URL's, rather than the named sources from the
chocolatey.config file, i.e. https://community.chocolatey.org/api/v2/
rather than simply "chocolatey".

This is important, since the full URL is what is used to ensure that a
"choco install all" is not being attempted against one of the
configured public sources.

This wasn't picked up during the initial PR review, but it was picked
up during manual testing, and this commit is the output of that manual
testing.