(#310) Fix path to chocolatey.lib files
When progressing the release of 2.3.0, it was found that the
chocolatey.lib nupkg didn't contain any files. This was due to a change
in the folder structure which wasn't caught during testing.

This commit fixes this folder structure, pointing directly to the
chocolatey.lib folder, so that running nuget pack works as expected.

There may be additional changes required to this portion of the build,
to make sure that everything is being created in the same way, but that
can be picked up post release/