(#3465) Build MSI on GHA
As with the previous commit, this adds the configuration for GitHub
Actions. Thus allowing us to ensure the MSI is able to be built in
GitHub Actions as well.

When running the builds in CI, we should generate the MSI as well. This
will allow us to get an MSI for any build to test with if needed.