(build) Add installation of WiX as a tool
While this is not strictly needed, as the MSI for Chocolatey CLI is not
always built, the inclusion of the *.wixproj in the Visual Studio *.sln
file means that some parts of the build (like the restore task), does
attempt to do stuff with WiX.  While this doesn't actually stop the
build from happening, as error is shown, and can cause confusion. Always
installing the WiX tool prevents these errors from showing.