(#3451) Use availablePackages in GetPackageDependencies
This updates the dependency handling of packages to use
availablePackages when a package from the same source satisfies the
dependency.  This improves packages that have dependencies with many
versions available, and allows less data being requested from servers
in the case we have already have available packages from the source.

Without this fix, the installation will be delayed with repeated
resolution of the same package dependencies.

Use parentPackage version when finding the parent package.