(doc) Update CREDITS.md file
During a review, it was found that the CREDITS.md file, which contains
information about all the 3rd party dependencies that Chocolatey CLI
takes, was not up to date.

This commit addresses this by updating all version numbers, both in the
table of contents, and the main body of the document, to match what is
actually being used.  In addition, the license that is in place for the
Microsoft.Web.Xdt dependency is updated to reflect that correct one, as
this was recently updated.

Finally, a small typo was fixed:
Licensed Terms -> License Terms