(maint) Add recommended extension
This extension is useful when editing Markdown files, especially when
it contains a Table Of Contents. This extension will update the TOC
simply by making changes to the main body.

In order to get this new file to be included, it was necessary to add
another exclusion to the rules in the .gitignore file.