(#3451) Add pester tests for slow dependency
This commit adds pester tests and files required for the pester tests.
The tests include ensuring that installing a package that contains a
dependency tree that spans lots of versions does not take unnecessarily
long. As well as duplicating an integration test into the pester tests
to provide us with some belt and suspenders to help us catch issues at
both the integration and the pester test levels.