(#3489) Remove usage of hard-coded port
In the previous commit, an attempt was made to remove the usage of the
hard-coded WireMock.Net port, however, not all usages of the 24626 port
were removed.

These have been replaced with .FormatWith(_wireMockServer.Url) where
needed, to make the port number dynamic based on what it is started
with.

This change is an extension of the comment that was left here:

https://github.com/chocolatey/choco/pull/3500#issuecomment-2308162787