(maint) Move these tests to WindowsOnly
Due to the fact that as part of listing packages, we now include the
hydration of remembered arguments, these tests no longer work on
anything but Windows. We may want to look into moving the hydration of
these remembered arguments to a different location, so that it doesn't
cause these types of problems.  In addition, it takes time to find and
hydrate this information, so if it isn't required, we may want to look
into only doing it when required.