(maint) Output choco pack failures
This adds logic to the Invoke-Tests file to output if a package failed
fail to pack. This allows for improved developer experience by
indicating if packages fail to build. This also allows impromptu
improvement to the developer experience by allowing simple changes to
the script to throw in place of outputting an errorand thus not
allowing tests to run without the full set of test packages.