(doc) Add testing section to link to other docs
It was pointed out in community discussion that the main README file
doesn't mention anything to do with running tests for the Chocolatey
CLI codebase.  This commit adds a new Testing section to the README.md
file, that links off to the pertinent information in the TESTING.md
file.

Used the Markdown All in One VSCode extension to update the table of
contents.  Initially this was adding an entry in the TOC section for the
first heading, which was not there before, so added an entry in
workspace settings.json file to exclude this.  Updating this table of
contents actually picked up on some previously entries which hadn't
been added.