(#3427) Remove use of -unpackself from the build
This command is deprecated, so we need to remove it from our build
to give ourselves the ability to actually remove that functionality in
future.

To facilitate this, the embedded resource files are being copied to the
build output folder so that the Cake build can easily copy them to the
locations they are needed.

In the longer-term these resource files will be relegate to `<None>`
rather than `<EmbeddedResource>` when we fully remove the `-unpackself`
command from Chocolatey CLI, and we will not need to be shipping
them embedded in the executable itself any longer.