(#3503, #3513) Better handle argument decryption
Currently we are decrypting the arguments file every time a list is run,
even though we only care about them in scenarios where we might output
them to the user (when Verbose is specified on the list command or with
the info --local-only command). This updates the handling of the
remembered arguments to be consistent between the List method and the
scenario in an Upgrade where we want to use the remembered arguments. We
now will only throw an error in a scenario where we can't decrypt the
contents and we need them for the scenario to succeed. We are also
stating what file we couldn't decrypt so the user can troubleshoot.