(#3534) Fix ILMerge of no7zip version of choco.exe
It was noted that during the compilation of the final release build,
the generated choco.exe for the no7zip version of Chocolatey CLI was
bigger then expected.

After investigation, it was found that tests project output, including
all the WireMock assemblies were being added to the output directory.
Due to the fact that the selection of the files to be ILMerge'd collects
all the files in the output directory, all of these files were being
put into the final choco.exe.

This commit addresses this problem by updating the build configuration
to remove the compiling of the test projects when it is the
ReleaseOfficialNo7zip configuration.  With this change in place, the
compiled choco.exe for the no7zip version is now smaller than the
normal choco.exe.