(#3526) Implement fallback logic from benchmarks
The primary method is quickest, but as it is an unstable API, we have a
need to provide a fallback that will be able to be used if the DLL is
missing/removed, or if the entry point is later removed.

Of the stable options, this p/invoke method seems to be the next best
option to work with.

Crucially, these p/invokes need to be provided by different types, so
that if there is a TypeLoadException from a failed p/invoke, we can
still attempt the fallback.

Also, added a couple more exclusions for the choco.exe shim and for
winlogon, which are not useful to include.