(#3526) Minor fixes for mono in the benchmarks and tests
RuntimeInformation is shadowed by Mono which means that build breaks and
because all of the namespaces are the same under Mono's version, that
name is not usable in our build currently.

We can reuse our existing Platform.GetPlatform() helper here instead,
rather than trying to disentangle exactly why this build configuration
is not particularly functional.