(#3364) Remove unusable tab completion tests
These tests are verifying the old method and are targeting an
implementation detail, rather than actually checking the tab completion
results, which the following tests do anyway.

As such, we're best off removing them as there is no direct way to check
the same things for the Register-ArgumentCompleter method. The tests for
the tab completion results themselves cover this adequately already.