(#3526) Remove test for an empty debug log
The debug log is used in a lot of places when we don't want the user to
see the output, but we want it available. At the moment it is not
consistent about what we do when they select debug and limit output. It
has been decided at the moment to remove these tests as they are now
breaking by us emitting the user agent to the debug log. A follow up
issue #3555 has been created to determine what we want for behaviour
when limit output and debug is selected.