(#3526) Fix user-agent tests for licensed
Licensed does queries in the background before / while normal commands
are running, so we can end up with multiple user-agent debug lines.

Resolve this by just looking at the most recently-outputted user agent,
as these should all look the same (and if not, the last one should be
the one from the actual command run).