(tests) Fix tests for new signing
This commit fixes hanging tests and bring the changes from 981a5d9 into
the support branch.

Going forward, the official signing certificate will only be in play on
tagged builds.  As such, we can't easily assert on the certificate
thumbprint. Instead, let's check on the Issuer name, and if the
official one, check for certificate being valid, otherwise let's assert
on it not being correct.  This is similar to what is now done elsewhere
when checking for signatures, i.e. in CCR and Chocolatey.Cake.Recipe.