(#3603) Don't allow trace logging when no elevated
Prior to this change, trace level logging was available to everyone.
However, due to the sensitive nature of some of the output, the
decision has been taken to restrict trace logging to only elevated
sessions.

When an attempt is made to use trace logging in a non-elevated session,
a warning will be shown, and no trace logging will be shown.  In
addition, if the -r option is in play, the warning about no trace
logging will go to the log file, but won't be displayed.