(#3618) Ensure ILMerge can run
Previously in #3609 we removed the No7zip build configuration. This
build configuration set itself as a dependee of ILMerge, and a Dependent
of Build and Test. This resulted in the ILMerge task being dependent
upon Build and Test.

This commit sets the ILMerge task to be dependent upon both Build and
Test so that builds that target just the ILMerge can succeed.