(maint) Remove unused dependency
The System.Linq.Dynamic.Core package is actually not being used, so it
is safe to remove from the project.  This came up as a Depdendabot
security report, which prompted us to check whether this was being
used.  When it was found that it isn't, the decision was made to remove
it, rather than updating it.