(#3621) Skip trace warning when trace not used
This updates the warning that is outputted in non-elevated mode to not
be outputted when the user has not specified `--trace` on the
commandline.

Without this change, any time a use are running in a non-elevated mode,
they will always have the trace warning outputted even when they haven't
used the argument.