(#3591) Remove no7zip build tasks
This commit removes the no longer required no7zip build tasks, and also
the associated sln and csproj entries.

Previously, it was necessary to build a version of choco.exe that didn't
include the 7zip binaries, however, this version of Chocolatey CLI is no
longer required.

To simplify the overall build process, and also the release steps
required, the decision was taken to completely remove the no7zip build.