(#3668) Add choco support command
This commit introduces a new choco support command for Chocolatey CLI.
This command already exists for Chocolatey Licensed Extension, but the
decision was made to also introduce it into open-source, for
consistency reasons.  The output from this version of the command is
different to what is shown when you have a license in play, as it
points the user to the open-source resources, rather than to reach out
to the support team.

Also included is some unit tests and some pester tests, to ensure that
the command is working as expected.