(#3344) Convert Get-PackageScriptParameter to helper cmdlet
It was pointed out that this PR should be targeted
at including new helper functions into the C# conversion
efforts.

This converts Get-PackageParameter to use a
helper class (shared between this and the new
cmdlet) and adds Get-PackageScriptParameters to the
Chocolatey.PowerShell compiled module.

Splits Get-functionality into private methods

Also refactors the helper method for package parameters to enable logging of sensitive parameter names (not values).