(#3294) Handle "|" in environment variable
When an environment variable contains a "|" character, during the
process of hydrating the values, the "|" causes an error, as it is
attempting to do "something" when executing the "set" command, from the
_env.cmd file.  We suspect that the underlying problem is the
combination of the " which is added to the environment variable by
Windows itself, and usage of the | as a logical OR operation.  To get
around this problem, we needed to escape the pipe character using a ^
character.  This is done using a replacement of a single "|" with "^|".

This commit adds this replacement, in addition to moving the endlocal
further down the file, so that the EnableDelayedExpansion is in place
when the operation is done.