(#1713) Previous package version as env variable
This commit adds the ability to output the previous package version,
during say a `choco upgrade` command, as an environment variable.
Prior to this commit, it was only the package version that was
currently being installed/upgraded that was presented as an environment
variable.

The method that is responsible for writing the environment variables is
the PreparePowerShellEnvironment method, which is defined by the
IPowershellService interface.  Rather than make a change to the
interface to pass in this information, it was decided to add a property
to the ChocolateyConfiguration class, which is currently being passed
through to this method.  This means that there is some maintainability
in place, where additional information can be passed through as
environment variables in the future, should the need arise.  This would
be done by simply adding another item into the Dictionary.

Pester tests have been updated to allow the testing for this new
environment variable during the process of upgrading a package version.

This is needed since there are times that during an upgrade you want to
do something based on which version is being upgraded from. For
example, you might have to do "something" when upgrading from version
1.0.0 and "something else" when upgrading from version 2.0.0.  Being
able to inspect the previous package version as an environment
variable, in the same way as you can the package version being
currently installed is _very_ useful, and also consistent with how
package maintainers get other information.