(#3700) Alphabetize all the things
This commit attempts to apply some clear organisation to the
ChocolateyTabExpansion.ps1 file by using alphabetical orderding, both in
terms of the commands that are listed out, as well as the options that
are passed into commands.  This will hopefully make this file easier to
update, and to see what options are applied where.