(#3666) Add ChocolateyNotSilent env variable
This commit introduces a new ChocolateyNotSilent environment variable,
which will be set when the user makes use of the --not-silent command
line argument.

In addition, a new Pester test assertion has been added to check for the
presence of the environment variable.

This is useful, as it will allow package maintainers to take additional
actions in their scripts, when required.  An example of this would be
when a package uses an autohotkey script to perform the installation.
Passing the --not-silent argument here does nothing, and the autohotkey
script is still ran.  The presence of this environment variable can now
be used as a trigger for "if" the autohotkey should be run.