(#3700) Update all commands and options
Using a newly created spreadsheet which lists all the
command/subcommands/options for Chocolatey CLI, update the
ChocolateyTabExpansion.ps1 file to use this "source of truth". This
includes adding missing commands/subcommands/options where required. In
addition, the new default option for each command is applied, and these
will be used elsewhere going forward, i.e. in our documentation, so that
we can start to get some consistency.

In addition, changes were made to the settings.json file for VSCode, so
that known words would not cause "Problems" to be reported when editting
the file.