(#3718) Add ability to always ignore HTTP cache
This commit adds the ability to enable a new alwaysIgnoreHttpCahe
feature, which does the same thing as using the '--ignore-http-cache'
option, just allowing it to _always_ be set that way, rather than having
to apply it to each command.

This has been a requested feature from the Chocolatey Community for a
while now.

Integration tests have been added to exercise the new feature to always
ignore the HTTP Cache.  The assertions in the tests are basically
verifying that there are either fewer or more HTTP requests based on
whether the feature is enabled or disabled.