(#3717) Add '--ignore-pinned' option to choco list
This commit makes it possible to list all the currently installed
packages, but ignore (i.e. skip over listing) of any packages that are
currently pinned.  This can be useful information to have when deciding
what operations can/should be performed on the currently installed
packages.  Without this option in place, it would have been necessary
to combine the output from choco list and choco pin list, to get this
information.  Now, this is much simpler by running:
choco list--ignore-pinned.

This commit also adds the needed unit tests to verify that the new
option appears in the OptionSet for the ChocolateyListCommand, as well
as extending the Pester tests to verify that when listing packages,
using the --ignore-pinned option means that pinned packages are no
longer shown.