(#2591) Fix broken Pester tests
We found an issue when trying to pass multiple arguments through to
Invoke-Choco. After a few different attempts to get things working, we
settled on explicitly creating an array of arguments, which eant a minor
change in Invoke-Choco itself to handle this correctly.

The logic for skipping the `choco license` test was simplified to only
include that command _when_ licensed is in play, rather than trying to
skip that test when license is in play.

Also added the required setup to allow execution of the required
commands, i.e. install an older version of a package to allow `choco
outdated` to work, and setting an apikey so the headers appear when
doing `choco apikey`.

Took the opportunity to fix a spelling mistake that was in nother file
as well:

encryped -> encrypted