(#669) Improve PowerShell Tab Completion
This commit improves the PowerShell Tab Completion for Chocolatey CLI,
by including completions for the following commands:

* apikey
* config
* feature
* pin
* rule
* source
* template

Each of these completions rely on directly calling the equivalent
"list£ sub-command for each command (with the exception of one of the
pin completions, which also relies on running choco list), which
unfortunately can take a second to complete.

As a future enhancement to this functionality, we should look to
serializing these completions to disk, so that they can be loaded and
shown from there. These files would be updated each time an associated
command is executed. For example, when someone runs "choco source add"
update the completions file for when doing "choco source remove".

A group of Pester tests have been added to ensure that these
completions work as expected.