(maint) Update Chocolatey CLI sln file
For a while now, opening the main Chocolatey CLI sln file in Visual
Studio has been causing an immediate modification to the file.  Rather
than just blindly accept the changes, as we are not clear on why they
are suddenly happening, Kim and I sat and went through the changes to
make sure they are valid.  We made sure that for the Wix configuration,
only the chocolatey.install project was selected (which is the whole
point of that configuration).  This resulted in a new collection of
additions to the sln file.