(#3694) Fix bug with case sensitive comparison
When checking for the existence of a specific template, the comparison
that was previously being done was a case sensitive one.  As a result,
if you have a template with the folder name of "NewFileInstaller" and
you search for it with a name of "newfileinstaller" it would not be
found, and you would get an error of "Unable to find requested
template".

This commit addresses this by ensuring to do a case insensitive
comparison, by first making sure that everything to set to lower
invariants of all the letters.  This is consistent with how other
things are being done in this file.  Ideally, we would make use of a
StringComparer, and do an insensitive search using that, however, this
is not available on the types that we are using.