(#3694) Fix edge case with template naming
In the previous commit associated with this issue:

https://github.com/chocolatey/choco/pull/3732/commits/1d1b180355ff20384b639bfa3b18ea6e212e9e7b

A change was made to try to make sure that everything related to
template, the name, the directories, etc are all lower case.  That way,
in the absence of a StringComparer, a simple comparison can be made.
However, one of our automated tests caught a problem with this approach
when the code is checking for the defaultTemplate and whether it had
been overridden or not.  This code path hadn't been ToLower'ed and as a
result, when there are uppercase letters in the path, for example the
"C:\ProgramData" location, the check would fail, and it would no longer
output that the default template had been overridden.

This commit addresses that problem by first ToLower'ing these values
ahead of the check.  It also switches to using the case-insensitive
IsEqualTo method, which makes the code slightly less complicated, and
easier to follow.